+++ src/include/utils/timestamp.h
@@ -98,8 +98,8 @@ extern PGDLLIMPORT TimestampTz PgReloadTime;
 extern TimestampTz GetCurrentTimestamp(void);
 extern void TimestampDifference(TimestampTz start_time, TimestampTz stop_time,
 								long *secs, int *microsecs);
-extern long TimestampDifferenceMilliseconds(TimestampTz start_time,
-											TimestampTz stop_time);
+extern int TimestampDifferenceMilliseconds(TimestampTz start_time,
+										   TimestampTz stop_time);
 extern bool TimestampDifferenceExceeds(TimestampTz start_time,
 									   TimestampTz stop_time,
 									   int msec);
